agno_example/pyproject.toml
[project]
name = "agno-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.4.0",
    "create-agent-app",
    "google-genai>=1.11.0",
    "langwatch-scenario>=0.1.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
create-agent-app = { path = "../", editable = true }
